jammy: move some things around to prepare for more versions
This is just shifting some bits raound to make more sense for adding
other Ubuntu versions.
