+++ cmd/frontend/main.go
@@ -7,7 +7,7 @@ import (
 	"github.com/Azure/dalec/frontend"
 	"github.com/Azure/dalec/frontend/azlinux"
 	"github.com/Azure/dalec/frontend/debug"
-	"github.com/Azure/dalec/frontend/jammy"
+	"github.com/Azure/dalec/frontend/ubuntu"
 	"github.com/Azure/dalec/frontend/windows"
 	"github.com/moby/buildkit/frontend/gateway/grpcclient"
 	"github.com/moby/buildkit/util/appcontext"
@@ -35,7 +35,7 @@ func main() {
 		frontend.WithBuiltinHandler(azlinux.Mariner2TargetKey, azlinux.NewMariner2Handler()),
 		frontend.WithBuiltinHandler(azlinux.AzLinux3TargetKey, azlinux.NewAzlinux3Handler()),
 		frontend.WithBuiltinHandler(windows.DefaultTargetKey, windows.Handle),
-		frontend.WithBuiltinHandler(jammy.DefaultTargetKey, jammy.Handle),
+		ubuntu.Handlers,
 		frontend.WithTargetForwardingHandler,
 	)); err != nil {
 		bklog.L.WithError(err).Fatal("error running frontend")
